Fixed travis
